refactor: Move all init logic into common abstract base class
